aider: Handled KeyError for 'descriptionSnippet' key in video dictionary.
# Context:
USER: KeyError: 'descriptionSnippet' in line 22
ASSISTANT: The KeyError indicates that the 'descriptionSnippet' key is not present in the video dictionary for some videos. This could be because some videos do not have a description.

To handle this, you can use a try-except block to catch the KeyError and assign a default value when the key is not present. Here's how you can modify your code:

```python
pages/video_list.py
<<<<<<< ORIGINAL
        description = video["descriptionSnippet"]["runs"][0]["text"]
=======
        try:
            description = video["descriptionSnippet"]["runs"][0]["text"]
        except KeyError:
            description = "No description"
>>>>>>> UPDATED
```

This will assign the value "No description" to the description variable when the 'descriptionSnippet' key is not present in the video dictionary.